requirements.txt
annotated-types==0.7.0
argcomplete==3.5.1
arrow==1.3.0
asgiref==3.8.0
b2==4.1.0
b2sdk==2.5.1
boto3==1.35.40
botocore==1.35.40
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
Django==5.0.3
django-role-permissions==3.2.0
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
jmespath==1.0.1
logfury==1.0.1
mysqlclient==2.2.4
packaging==24.1
phx-class-registry==4.1.0
pillow==10.3.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rst2ansi==0.1.5
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
tqdm==4.66.5
types-python-dateutil==2.9.0.20241003
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
djangorestframework-simplejwt
